requirements.txt
fastapi==0.112.0
func_timeout==4.3.5
httpx==0.27.0
launart==0.8.2
Pillow==10.4.0
pluginlib==0.9.2
pydantic==2.8.2
PyYAML==6.0.2
Requests==2.32.3
satori==0.1.5
schedule==1.2.2
sympy==1.13.2
urllib3==2.2.2
uvicorn==0.30.5
xmltodict==0.13.0
